refactor(utils): Split the utility function
The `getGitHubUrl()` is not an env-related function.
