+++ src/libs/env.ts
@@ -1,5 +1,3 @@
-import { join } from "@std/path";
-
 import type { Repository } from "./types.ts";
 
 export function getRepository(): Repository {
@@ -19,19 +17,3 @@ export function getRepository(): Repository {
 }
 
 export const githubToken: string | undefined = Deno.env.get("GITHUB_TOKEN");
-
-export function getGitHubUrl(
-  repository: Repository,
-  ref: string = "master",
-): URL {
-  return new URL(
-    join(
-      repository.owner,
-      repository.name,
-      "blob",
-      ref,
-      repository.path,
-    ),
-    "https://github.com",
-  );
-}
